// src/interFace/type.ts
export type TCounter = {
  id: number;
  cunterNum: number;
  image?: StaticImageData;
  skillName?: string;
  bgColor?: string;
  wowDelay?: string;
  text?: string;
  icon?: string;
  desc?: string
}